# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 Swagger API",
        "description": "L5 Swagger API description",
        "contact": {
            "email": "saitokdev@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/conversation/user/unread": {
            "get": {
                "tags": [
                    "Conversation"
                ],
                "summary": "Get unread conversations",
                "operationId": "e8665d890367807b10cda5c1ee7b2575",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/conversation/all/{page}": {
            "get": {
                "tags": [
                    "Conversation"
                ],
                "summary": "Get conversations",
                "operationId": "9764bc24747bc406628c5e81565c0304",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/page"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/conversation/message/{conversationId}/{page}": {
            "get": {
                "tags": [
                    "Conversation"
                ],
                "summary": "Get messages by conversation Id",
                "operationId": "acac3eb042d537fae0d88e9c22aaeae1",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/conversationId"
                    },
                    {
                        "$ref": "#/components/parameters/page"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/conversation/message/add": {
            "post": {
                "tags": [
                    "Conversation"
                ],
                "summary": "Add message to the conversation",
                "operationId": "48adafdf9b34b48835d8739c859b4524",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "conversationId": {
                                        "type": "integer"
                                    },
                                    "conversationName": {
                                        "type": "string"
                                    },
                                    "body": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/conversation/message/{id}": {
            "get": {
                "tags": [
                    "Conversation"
                ],
                "summary": "Get message by Id",
                "operationId": "4fb8fe7ada1eb7a5490f0cc0aae8e846",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/messageId"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/conversation/message/get": {
            "post": {
                "tags": [
                    "Conversation"
                ],
                "summary": "Get message by Ids",
                "operationId": "c319570bd62fecb90e2b193934b06b75",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "type": "integer"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/conversation/get/{id}": {
            "get": {
                "tags": [
                    "Conversation"
                ],
                "summary": "Get conversation by Id",
                "operationId": "0eb2a4bf7758850891607eead0440bd2",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/conversation_id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/user/device/add": {
            "post": {
                "tags": [
                    "Device"
                ],
                "summary": "Add device to the user",
                "operationId": "ab9c210098c89785a9a0584e144c94b2",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "token": {
                                        "type": "string"
                                    },
                                    "type": {
                                        "type": "string"
                                    },
                                    "appVersion": {
                                        "type": "string"
                                    },
                                    "model": {
                                        "type": "string"
                                    },
                                    "osVersion": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "parameters": {
...
            "conversationId": {
                "name": "conversationId",
                "in": "path",
                "allowReserved": true,
                "schema": {
                    "type": "integer"
                }
            },
            "messageId": {
                "name": "messageId",
                "in": "path",
                "allowReserved": true,
                "schema": {
                    "type": "integer"
                }
            },
            "conversation_id": {
                "name": "id",
                "in": "path",
                "allowReserved": true,
                "schema": {
                    "type": "integer"
                }
            },
...
            "page": {
                "name": "page",
                "in": "path",
                "required": true,
                "schema": {
                    "type": "integer"
                }
            },
...
        },
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}